[Gepardec/mega#699] split endpoint into service and call service in endpoint, add schedule
