+++ src/main/java/com/gepardec/mega/service/api/SyncService.java
@@ -1,5 +1,9 @@
 package com.gepardec.mega.service.api;
 
+import com.gepardec.mega.rest.model.EmployeeDto;
+
+import java.util.List;
+
 /**
  * @author Thomas Herzog <herzog.thomas81@gmail.com>
  * @since 10/3/2020
@@ -7,5 +11,6 @@ package com.gepardec.mega.service.api;
 public interface SyncService {
 
     void syncEmployees();
+    List<EmployeeDto> syncUpdateEmployeesWithoutTimeBookingsAndAbsentWholeMonth();
 
 }
