+++ src/main/java/com/gepardec/mega/service/impl/SyncServiceImpl.java
@@ -1,15 +1,21 @@
 package com.gepardec.mega.service.impl;
 
 import com.gepardec.mega.application.configuration.ApplicationConfig;
+import com.gepardec.mega.db.entity.common.AbsenceType;
 import com.gepardec.mega.db.entity.employee.User;
 import com.gepardec.mega.db.repository.UserRepository;
 import com.gepardec.mega.domain.model.Employee;
 import com.gepardec.mega.domain.model.Project;
 import com.gepardec.mega.domain.utils.DateUtils;
+import com.gepardec.mega.notification.mail.dates.OfficeCalendarUtil;
+import com.gepardec.mega.rest.mapper.EmployeeMapper;
+import com.gepardec.mega.rest.model.EmployeeDto;
 import com.gepardec.mega.service.api.EmployeeService;
 import com.gepardec.mega.service.api.ProjectService;
 import com.gepardec.mega.service.api.SyncService;
 import com.gepardec.mega.service.mapper.SyncServiceMapper;
+import com.gepardec.mega.zep.ZepService;
+import de.provantis.zep.FehlzeitType;
 import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
@@ -17,6 +23,8 @@ import org.apache.commons.lang3.time.StopWatch;
 import org.slf4j.Logger;
 
 import java.time.Instant;
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -38,6 +46,9 @@ public class SyncServiceImpl implements SyncService {
     EmployeeService employeeService;
 
     @Inject
+    ZepService zepService;
+
+    @Inject
     ProjectService projectService;
 
     @Inject
@@ -48,6 +59,12 @@ public class SyncServiceImpl implements SyncService {
 
     @Inject
     SyncServiceMapper mapper;
+
+    @Inject
+    EmployeeMapper employeeMapper;
+
+    @Inject
+    Logger logger;
 
     @Override
     public void syncEmployees() {
@@ -70,6 +87,69 @@ public class SyncServiceImpl implements SyncService {
 
         log.info("User sync took: {}ms", stopWatch.getTime());
         log.info("Finished user sync: {}", Instant.ofEpochMilli(stopWatch.getStartTime() + stopWatch.getTime()));
+    }
+
+    @Override
+    public List<EmployeeDto> syncUpdateEmployeesWithoutTimeBookingsAndAbsentWholeMonth() {
+        //to avoid having a look at external employees filter
+        List<Employee> activeAndInternalEmployees = employeeService.getAllActiveEmployees()
+                .stream()
+                .filter(e -> !e.getUserId().startsWith("e"))
+                .toList();
+        List<EmployeeDto> updatedEmployees = new ArrayList<>();
+        List<Employee> absentEmployees = new ArrayList<>();
+
+        LocalDate now = LocalDate.now();
+        LocalDate firstOfPreviousMonth = now.withMonth(now.getMonth().minus(1).getValue()).withDayOfMonth(1);
+        //use this firstOfPreviousMonth.getYear() because of january and december
+        LocalDate lastOfPreviousMonth = DateUtils.getLastDayOfMonth(firstOfPreviousMonth.getYear(), firstOfPreviousMonth.getMonth().getValue());
+
+        for (var employee : activeAndInternalEmployees) {
+            //considering all absence types besides HomeOffice and External training days
+            List<FehlzeitType> absences = zepService.getAbsenceForEmployee(employee, firstOfPreviousMonth).stream()
+                    .filter(absence -> !AbsenceType.getAbscenceTypesWhereWorkingTimeNeeded().stream()
+                            .map(AbsenceType::getAbsenceName).toList()
+                            .contains(absence.getFehlgrund()))
+                    .toList();
+            boolean allAbsent = true;
+
+            for (LocalDate day = firstOfPreviousMonth; !day.isAfter(lastOfPreviousMonth); day = day.plusDays(1)) {
+                if (OfficeCalendarUtil.isWorkingDay(day)) {
+                    boolean isAbsent = isAbsent(day, absences);
+                    if (!isAbsent) {
+                        allAbsent = false;
+                        break;
+                    }
+                }
+            }
+
+            // only add employee who was absent the whole month
+            if(allAbsent){
+                absentEmployees.add(employee);
+            }
+        }
+
+        // set release date of employee to last day of previous month --> no confirmation of employee necessary
+        absentEmployees.forEach(employee -> {
+            zepService.updateEmployeesReleaseDate(employee.getUserId(), lastOfPreviousMonth.toString());
+            updatedEmployees.add(employeeMapper.mapToDto(zepService.getEmployee(employee.getUserId())));
+        });
+
+        logger.info("updated " + updatedEmployees.size() + " employee(s)!");
+        return updatedEmployees;
+    }
+
+    private boolean isAbsent(LocalDate day, List<FehlzeitType> absences) {
+        for(var absence : absences){
+            LocalDate startDate = LocalDate.parse(absence.getStartdatum());
+            LocalDate endDate = LocalDate.parse(absence.getEnddatum());
+            if(day.equals(startDate) ||
+                    day.equals(endDate) ||
+                    (day.isAfter(startDate) && day.isBefore(endDate))) {
+                return true;
+            }
+        }
+        return false;
     }
 
     private void createNonExistentUsers(final List<Employee> employees, final List<User> users, final List<Project> projects) {
